+++ examples/editor/src/App.jsx
@@ -6,7 +6,7 @@ import { convertFromRaw, convertToRaw, EditorState } from '@wix/draft-js';
 import Plugins from './Plugins';
 import ModalsMap from './ModalsMap';
 import { EditorState as RichEditorState, RichContentEditor, RichContentEditorModal } from 'wix-rich-content-editor';
-import { Button, normalizeInitialState } from 'wix-rich-content-common';
+import { Button, normalizeInitialState, TOOLBARS } from 'wix-rich-content-common';
 import { testImages, testVideos } from './mock';
 import './App.css';
 import theme from './theme/theme'; // must import after custom styles
@@ -68,15 +68,75 @@ class App extends Component {
             250),
         ),
       },
-      // image: {
-      // imageProps: src => ({
-      //   'data-pin-media': `https://static.wixstatic.com/media/${src.file_name}`,
-      //   'data-pin-url': 'https://wix.com',
-      // }),
-      // toolbar: {
-      //   hidden: ['link', 'settings']
-      // }
-      // },
+      // getToolbarSettings: ({ pluginButtons, textButtons }) => [
+      //   {
+      //     name: TOOLBARS.SIDE,
+      //     getVisibilityFn: () => ({
+      //       desktop: () => true,
+      //       mobile: {
+      //         ios: () => true,
+      //         android: () => true,
+      //       }
+      //     }),
+      //   },
+      //   {
+      //     name: TOOLBARS.MOBILE,
+      //     shouldCreate: () => ({
+      //       desktop: false,
+      //       mobile: {
+      //         ios: false,
+      //         android: false,
+      //       }
+      //     }),
+      //   },
+      //   {
+      //     name: TOOLBARS.FOOTER,
+      //     getPositionOffset: () => ({
+      //       mobile: {
+      //         ios: { x: 0, y: 500 },
+      //       }
+      //     }),
+      //     getVisibilityFn: () => ({
+      //       desktop: () => false,
+      //       mobile: {
+      //         ios: () => true,
+      //         android: () => true,
+      //       }
+      //     }),
+      //     shouldCreate: () => ({
+      //       desktop: () => false,
+      //       mobile: {
+      //         ios: () => true,
+      //         android: () => true,
+      //       }
+      //     }),
+      //     getButtons: () => ({
+      //       desktop: () => [],
+      //       mobile: {
+      //         ios: pluginButtons.filter(({ buttonSettings }) => buttonSettings.toolbars.includes(TOOLBARS.FOOTER))
+      //         .map(({ component }) => component),
+      //         android: () => [],
+      //       }
+      //     }),
+      //   },
+      //   {
+      //     name: TOOLBARS.STATIC,
+      //     getPositionOffset: () => ({
+      //       desktop: { x: 40, y: 40 },
+      //       mobile: {
+      //         ios: { x: 40, y: 40 },
+      //         android: { x: 50, y: 50 },
+      //       }
+      //     }),
+      //     getVisibilityFn: () => ({
+      //       desktop: () => false,
+      //       mobile: {
+      //         ios: () => true,
+      //         android: () => true,
+      //       }
+      //     }),
+      //   }
+      // ]
     };
     const mockUpload = (files, updateEntity) => {
       //mock upload