+++ packages/common/src/index.js
@@ -82,7 +82,8 @@ export {
 export { default as WixUtils } from './Utils/wixUtils';
 export { sizeClassName, alignmentClassName, textWrapClassName } from './Utils/classNameStrategies';
 export { getSelectionStyles } from './Utils/inlineStyleUtils';
+export { getConfigByFormFactor } from './Utils/getConfigByFormFactor';
 export { default as createInlineStyleDecorators } from './Decorators/inline-style-decorators';
 
 export { HEADING } from './Decorators/inline-styles';
-export { COMMANDS, MODIFIERS } from './consts';
+export { COMMANDS, MODIFIERS, TOOLBARS } from './consts';