+++ packages/plugin-divider/src/toolbar/insert-buttons.jsx
@@ -1,3 +1,4 @@
+import { TOOLBARS } from 'wix-rich-content-common';
 import { InsertPluginIcon } from '../icons';
 import { DEFAULTS } from '../constants';
 
@@ -6,6 +7,7 @@ export default ({ helpers, t }) => {
     {
       name: 'Divider',
       tooltipText: t('DividerPlugin_InsertButton_Tooltip'),
+      toolbars: [TOOLBARS.FOOTER, TOOLBARS.SIDE],
       Icon: InsertPluginIcon,
       componentData: DEFAULTS,
       helpers,