+++ packages/plugin-gallery/src/toolbar/insert-buttons.jsx
@@ -1,3 +1,4 @@
+import { TOOLBARS } from 'wix-rich-content-common';
 import { getDefault } from './../gallery-component';
 import { InsertPluginIcon } from '../icons';
 
@@ -11,6 +12,7 @@ export default ({ helpers, t }) => {
       Icon: InsertPluginIcon,
       componentData: getDefault(),
       helpers,
+      toolbars: [TOOLBARS.FOOTER, TOOLBARS.SIDE],
     },
   ];
 };